requirements.txt
torch>=2.0.0
torchvision>=0.15.0
nvidia-dali-cuda120>=1.28.0
webdataset>=0.2.86
ray[train]>=2.5.0
wandb>=0.15.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
tritonclient[http]>=2.34.0
pycryptodome>=3.18.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
